A data.frame containing the summary of differential markers for scEpigenomics dataset (Paired-Tag - H3K27ac)

A summary of differential markers for scEpigenomics (Paired-Tag - H3K27ac)

data("IDC_summary_scEpigenomics")

Format

IDC_summary_scEpigenomics - A data.frame obtained by iterative_differential_clustering().

References

Joint profiling of histone modifications and transcriptome in single cells from mouse brain, Chenxu Zhu, Yanxiao Zhang, Yang Eric Li, Jacinta Lucero, . Margarita Behrens, Bing Ren, Nature Methods, 2021 Paired-Tag