A Seurat object containing the scRNA part of Paired-Tag dataset. (Paired-Tag - scRNA)

A Seurat object of a subset of 2000 x 2000 cells by genes of the Paired-Tag multiomic paper. The metadata contains annotation from the author as well as iterative differential clusters run on the subset of cells and genes. The counts were analyzed with Seurat package.

data("Seu")

Format

Seu - A Seurat object. See Seurat::Seurat().

References

Joint profiling of histone modifications and transcriptome in single cells from mouse brain, Chenxu Zhu, Yanxiao Zhang, Yang Eric Li, Jacinta Lucero, . Margarita Behrens, Bing Ren, Nature Methods, 2021 Paired-Tag