Differential analysis with ChromSCape (differential activation)

Runs ChromSCape::differential_activation() function on clusters in the 'by' column of the SingleCellExperiment object and returns a data.frame containing the differential features that passed the thresholds.

differential_ChromSCape(
  object,
  by = "IDcluster",
  logFC.th = log2(1.5),
  qval.th = 0.01,
  min.pct = 0.01
)

Arguments

object

A SingleCellExperiment object containing scRNA dataset with a metadata column name matching the by parameter

by

A character specifying the name of the metadata column referencing the clusters.

logFC.th

A numeric specifying the log2 fold change of activation above/below which a feature is considered as significantly differential.

qval.th

A numeric specifying the adjusted p-value below which a feature is considered as significantly differential.

min.pct

Minimum percentage of cells to be active in the cells of the cluster to consider a feature as potentially significantly differential.

Value

See also

Examples

if(requireNamespace("ChromSCape", quietly=TRUE)){
 data("scExp", package = "IDclust")
DA <- differential_ChromSCape(scExp)
}