iterative_differential_clustering

Main function of the IDclust package. Provided a SingleCellExperiment pre-processed with ChromSCape, will find biologically relevant clusters by iteratively re-clustering and re-processing clusters. At each iteration, subclusters having enough significantly enriched features compared to other subclusters are defined as 'true' subclusters. Others are assigned to parent clusters. The algorithm will stop when no more 'true' subclusters are found.

This method ensure that each cluster found in this unsupervised way have significant biological differences, based on the user defined thresholds.

iterative_differential_clustering(object, ...)

# S3 method for default
iterative_differential_clustering(
  object,
  output_dir = "./",
  plotting = TRUE,
  saving = TRUE,
  n_dims = 10,
  dim_red = "PCA",
  vizualization_dim_red = "UMAP",
  processing_function = processing_ChromSCape,
  quantile.activation = 0.7,
  differential_function = differential_ChromSCape,
  logFC.th = log2(1.5),
  qval.th = 0.01,
  min_frac_cell_assigned = 0.1,
  limit = 5,
  starting.k = 100,
  starting.resolution = 0.1,
  resolution = 0.8,
  k = 50,
  limit_by_proportion = NULL,
  color = NULL,
  nThreads = 10,
  verbose = TRUE,
  ...
)

# S3 method for Seurat
iterative_differential_clustering(
  object,
  output_dir = "./",
  plotting = TRUE,
  saving = TRUE,
  n_dims = 50,
  dim_red = "pca",
  vizualization_dim_red = "umap",
  processing_function = processing_Seurat,
  differential_function = differential_edgeR_pseudobulk_LRT,
  logFC.th = log2(1.5),
  qval.th = 0.01,
  min_frac_cell_assigned = 0.1,
  limit = 10,
  starting.resolution = 0.1,
  starting.k = 100,
  resolution = 0.8,
  k = 100,
  color = NULL,
  nThreads = 10,
  verbose = TRUE,
  ...
)

Arguments

object
A Seurat object preprocessed with Seurat.

...
Additional parameters passed to the differential_function. See differential_edgeR_pseudobulk_LRT() for more information on additional parameters for the default function.

output_dir
The output directory in which to plot and save objects.

plotting
A logical specifying wether to save the plots or not.

saving
A logical specifying wether to save the data or not.

n_dims
An integer specifying the number of first dimensions to keep in the dimensionality reduction step.

dim_red
The name of the slot to save the dimensionality reduction at each step in the Seurat::Reductions(object).

vizualization_dim_red
The name of the slot used for plotting. Must be a valid slot present in Seurat::Reductions(object).

processing_function
A function that re-process the subset of clusters at each step. It msut take in entry a Seurat object, dim_red and n_dims parameters and returns a Seurat containing a cell embedding. See processing_Seurat for the default function.

quantile.activation
A numeric between 0 and 1 specifying the quantile of global activation to take as minimal percentage of activation for the differential analysis. Increasing this value will decrease the number of differential features.

differential_function
A function that take in entry a SingleCellExperiment object and parameters passed in ... and returns a data.frame containing the significantly differential features for each cluster. See differential_edgeR_pseudobulk_LRT for the default function.

logFC.th
A numeric specifying the log2 fold change of activation above/below which a feature is considered as significantly differential passed to the differential_function.

qval.th
A numeric specifying the adjusted p-value below which a feature is considered as significantly differential passed to the differential_function.

min_frac_cell_assigned
A numeric between 0 and 1 specifying the minimum percentage of the total cells in the SingleCellExperiment object that needs to be assigned. If a lower proportion is assigned, all cells are assigned to the cluster of origin.

limit
An integer specifying the minimum number of significantly enriched / depleted features required in order for a subcluster to be called a 'true' subcluster

starting.k
An integer specifying the number of nearest neighbors to use for the Louvain clustering of the first iteration. It is recommended to set it quite high in order to have few starting clusters

starting.resolution
A numeric specifying the resolution to use for the Louvain clustering of the first iteration. It is recommended to set it quite low in order to have few starting clusters.

resolution
A numeric specifying the resolution to use for the Louvain clustering at each iteration.

k
An integer specifying the number of nearest neighbors to use for the Louvain clustering at each iteration.

limit_by_proportion
Optional. A data.frame containing 3 columns - ncells, mean_n_differential, sd_n_differential - that reflect the number of false positive expected for a given cluster size.

color
Set of colors to use for the coloring of the clusters. This must contains enough colors for each cluster (minimum 20 colors, but 100 colors at least is recommended, based on the dataset).

nThreads
An integer specifying of threads to use for the calculation of the FDR.

verbose
A logical specifying wether to print.

Value

The SingleCellExperiment object with the assignation of cells to clusters. If saving is true, also saves list of differential analyses, differential analyses summaries and embeddings for each re-clustered cluster. If runFDR is TRUE, also saves the list of FDR for each re-clusterd cluster.

The Seurat object with the assignation of cells to clusters. If saving is true, also saves list of differential analyses, differential analyses summaries and embeddings for each re-clustered cluster.

Details

The default differential analysis used is the ChromSCape::differential_activation() function. This function compares the % of active cells in the cluster versus the rest of cells and perform a Chi-squared test to calculate p-values.

Examples

# Clustering of Seurat scRNA object (Paired-Tag)
if(requireNamespace("Seurat", quietly=TRUE)){

data("Seu", package = "IDclust")
Seu = iterative_differential_clustering(Seu, saving = FALSE, plotting =FALSE,
logFC.th = 0.2, qval.th = 0.1)

}
#> Initial round of clustering - limit of differential genes set to 0  for this first round only.
#> A1 - Found 28 differential regions.
#> A2 - Found 43 differential regions.
#> A3 - Found 29 differential regions.
#> Finished finding differences -  1  fraction of cells were assigned.
#> Doing iteration number  1 ...
#> Re-calculating PCA and subclustering for cluster Omega:A1 .
#> Warning: pseudoinverse used at -3.0999
#> Warning: neighborhood radius 0.4971
#> Warning: reciprocal condition number  1.673e-15
#> Warning: There are other near singularities as well. 0.090619
#> Warning: The following arguments are not used: use.dimred
#> Warning: The following arguments are not used: use.dimred
#> Warning: The following arguments are not used: dims
#> Warning: The following arguments are not used: dims
#> Found 5 subclusters.
#> clusters
#>  B1  B2  B3  B4  B5 
#> 420 356 166 136 124 
#> B3 - Found 20 differential regions.
#> B2 - Found 11 differential regions.
#> B1 - Found 5 differential regions.
#> B1  cluster has less than 10 enriched features.
#> Assigning the cells to cluster of origin.
#> B4 - Found 10 differential regions.
#> B5 - Found 15 differential regions.
#> Finished finding differences -  0.6505824  fraction of cells were assigned.
#> Doing iteration number  2 ...
#> Re-calculating PCA and subclustering for cluster Omega:A2 .
#> Warning: pseudoinverse used at -2.6255
#> Warning: neighborhood radius 0.31903
#> Warning: reciprocal condition number  1.2102e-29
#> Warning: There are other near singularities as well. 0.031008
#> Warning: The following arguments are not used: use.dimred
#> Warning: The following arguments are not used: use.dimred
#> Warning: The following arguments are not used: dims
#> Warning: The following arguments are not used: dims
#> Found 2 subclusters.
#> clusters
#>  B1  B2 
#> 248 157 
#> B1 - Found 84 differential regions.
#> B2 - Found 65 differential regions.
#> Finished finding differences -  1  fraction of cells were assigned.
#> Doing iteration number  3 ...
#> Re-calculating PCA and subclustering for cluster Omega:A3 .
#> Warning: pseudoinverse used at -2.6119
#> Warning: neighborhood radius 0.31851
#> Warning: reciprocal condition number  1.3293e-30
#> Warning: There are other near singularities as well. 0.031008
#> Warning: The following arguments are not used: use.dimred
#> Warning: The following arguments are not used: use.dimred
#> Warning: The following arguments are not used: dims
#> Warning: The following arguments are not used: dims
#> Found 3 subclusters.
#> clusters
#>  B1  B2  B3 
#> 183 140  70 
#> Not enough cells to form 2 replicates ... assigning 0 differential genes.
#> B1 - Found 25 differential regions.
#> B3 - Found 0 differential regions.
#> B3  cluster has less than 10 enriched features.
#> Assigning the cells to cluster of origin.
#> B2 - Found 17 differential regions.
#> Finished finding differences -  0.821883  fraction of cells were assigned.
#> Doing iteration number  4 ...
#> Re-calculating PCA and subclustering for cluster Omega:A1:B3 .
#> Warning: at  -2.2361
#> Warning: radius  0.0002545
#> Warning: all data on boundary of neighborhood. make span bigger
#> Warning: pseudoinverse used at -2.2361
#> Warning: neighborhood radius 0.015953
#> Warning: reciprocal condition number  1
#> Warning: There are other near singularities as well. 0.090619
#> Warning: zero-width neighborhood. make span bigger
#> Warning: The following arguments are not used: use.dimred
#> Warning: The following arguments are not used: use.dimred
#> Warning: The following arguments are not used: dims
#> Warning: The following arguments are not used: dims
#> Found only 1 subcluster for Omega:A1:B3 . Maximum clustering reached...
#> Doing iteration number  5 ...
#> Re-calculating PCA and subclustering for cluster Omega:A1:B2 .
#> Warning: pseudoinverse used at -2.5695
#> Warning: neighborhood radius 0.31909
#> Warning: reciprocal condition number  9.8978e-30
#> Warning: There are other near singularities as well. 0.031008
#> Warning: The following arguments are not used: use.dimred
#> Warning: The following arguments are not used: use.dimred
#> Warning: The following arguments are not used: dims
#> Warning: The following arguments are not used: dims
#> Found 2 subclusters.
#> clusters
#>  C1  C2 
#> 198 158 
#> C2 - Found 20 differential regions.
#> C1 - Found 17 differential regions.
#> Finished finding differences -  1  fraction of cells were assigned.
#> Doing iteration number  6 ...
#> Omega:A1 - This cluster was formed by 'unassigned' cells, not clustering it further...
#> Doing iteration number  7 ...
#> Re-calculating PCA and subclustering for cluster Omega:A1:B4 .
#> Warning: at  -2.1484
#> Warning: radius  0.00022215
#> Warning: all data on boundary of neighborhood. make span bigger
#> Warning: pseudoinverse used at -2.1484
#> Warning: neighborhood radius 0.014905
#> Warning: reciprocal condition number  1
#> Warning: There are other near singularities as well. 0.090619
#> Warning: zero-width neighborhood. make span bigger
#> Warning: The following arguments are not used: use.dimred
#> Warning: The following arguments are not used: use.dimred
#> Warning: The following arguments are not used: dims
#> Warning: The following arguments are not used: dims
#> Found only 1 subcluster for Omega:A1:B4 . Maximum clustering reached...
#> Doing iteration number  8 ...
#> Re-calculating PCA and subclustering for cluster Omega:A1:B5 .
#> Warning: at  -2.1094
#> Warning: radius  0.00025605
#> Warning: all data on boundary of neighborhood. make span bigger
#> Warning: pseudoinverse used at -2.1094
#> Warning: neighborhood radius 0.016002
#> Warning: reciprocal condition number  1
#> Warning: There are other near singularities as well. 0.090619
#> Warning: zero-width neighborhood. make span bigger
#> Warning: The following arguments are not used: use.dimred
#> Warning: The following arguments are not used: use.dimred
#> Warning: The following arguments are not used: dims
#> Warning: The following arguments are not used: dims
#> Found only 1 subcluster for Omega:A1:B5 . Maximum clustering reached...
#> Doing iteration number  9 ...
#> Re-calculating PCA and subclustering for cluster Omega:A2:B1 .
#> Warning: at  -2.4124
#> Warning: radius  0.00032181
#> Warning: all data on boundary of neighborhood. make span bigger
#> Warning: pseudoinverse used at -2.4124
#> Warning: neighborhood radius 0.017939
#> Warning: reciprocal condition number  1
#> Warning: There are other near singularities as well. 0.090619
#> Warning: zero-width neighborhood. make span bigger
#> Warning: The following arguments are not used: use.dimred
#> Warning: The following arguments are not used: use.dimred
#> Warning: The following arguments are not used: dims
#> Warning: The following arguments are not used: dims
#> Found only 1 subcluster for Omega:A2:B1 . Maximum clustering reached...
#> Doing iteration number  10 ...
#> Re-calculating PCA and subclustering for cluster Omega:A2:B2 .
#> Warning: at  -2.2112
#> Warning: radius  0.00023529
#> Warning: all data on boundary of neighborhood. make span bigger
#> Warning: pseudoinverse used at -2.2112
#> Warning: neighborhood radius 0.015339
#> Warning: reciprocal condition number  1
#> Warning: zero-width neighborhood. make span bigger
#> Warning: There are other near singularities as well. 0.090619
#> Warning: The following arguments are not used: use.dimred
#> Warning: The following arguments are not used: use.dimred
#> Warning: The following arguments are not used: dims
#> Warning: The following arguments are not used: dims
#> Found only 1 subcluster for Omega:A2:B2 . Maximum clustering reached...
#> Doing iteration number  11 ...
#> Re-calculating PCA and subclustering for cluster Omega:A3:B1 .
#> Warning: at  -2.2776
#> Warning: radius  0.00022986
#> Warning: all data on boundary of neighborhood. make span bigger
#> Warning: pseudoinverse used at -2.2776
#> Warning: neighborhood radius 0.015161
#> Warning: reciprocal condition number  1
#> Warning: There are other near singularities as well. 0.090619
#> Warning: zero-width neighborhood. make span bigger
#> Warning: The following arguments are not used: use.dimred
#> Warning: The following arguments are not used: use.dimred
#> Warning: The following arguments are not used: dims
#> Warning: The following arguments are not used: dims
#> Found only 1 subcluster for Omega:A3:B1 . Maximum clustering reached...
#> Doing iteration number  12 ...
#> Omega:A3 - This cluster was formed by 'unassigned' cells, not clustering it further...
#> Doing iteration number  13 ...
#> Re-calculating PCA and subclustering for cluster Omega:A3:B2 .
#> Warning: at  -2.1626
#> Warning: radius  0.00027006
#> Warning: all data on boundary of neighborhood. make span bigger
#> Warning: pseudoinverse used at -2.1626
#> Warning: neighborhood radius 0.016433
#> Warning: reciprocal condition number  1
#> Warning: There are other near singularities as well. 0.090619
#> Warning: zero-width neighborhood. make span bigger
#> Warning: The following arguments are not used: use.dimred
#> Warning: The following arguments are not used: use.dimred
#> Warning: The following arguments are not used: dims
#> Warning: The following arguments are not used: dims
#> Found only 1 subcluster for Omega:A3:B2 . Maximum clustering reached...
#> Doing iteration number  14 ...
#> Re-calculating PCA and subclustering for cluster Omega:A1:B2:C2 .
#> Warning: at  -2.2143
#> Warning: radius  0.00024543
#> Warning: all data on boundary of neighborhood. make span bigger
#> Warning: pseudoinverse used at -2.2143
#> Warning: neighborhood radius 0.015666
#> Warning: reciprocal condition number  1
#> Warning: There are other near singularities as well. 0.090619
#> Warning: zero-width neighborhood. make span bigger
#> Warning: The following arguments are not used: use.dimred
#> Warning: The following arguments are not used: use.dimred
#> Warning: The following arguments are not used: dims
#> Warning: The following arguments are not used: dims
#> Found only 1 subcluster for Omega:A1:B2:C2 . Maximum clustering reached...
#> Doing iteration number  15 ...
#> Re-calculating PCA and subclustering for cluster Omega:A1:B2:C1 .
#> Warning: at  -2.3139
#> Warning: radius  0.00029551
#> Warning: all data on boundary of neighborhood. make span bigger
#> Warning: pseudoinverse used at -2.3139
#> Warning: neighborhood radius 0.01719
#> Warning: reciprocal condition number  1
#> Warning: There are other near singularities as well. 0.090619
#> Warning: zero-width neighborhood. make span bigger
#> Warning: The following arguments are not used: use.dimred
#> Warning: The following arguments are not used: use.dimred
#> Warning: The following arguments are not used: dims
#> Warning: The following arguments are not used: dims
#> Found only 1 subcluster for Omega:A1:B2:C1 . Maximum clustering reached...
#> 
#> 
#> 
#> ##########################################################
#> Finished !
#> Found a total of 11 clusters after 15 iterations. 
#> The average cluster size is  181  and the median is 158 . 
#> The number of initital clusters not subclustered is  0 . 
#> ##########################################################

# Clustering of scExp scH3K27ac object (Paired-Tag)
if(requireNamespace("ChromSCape", quietly=TRUE)){

data("scExp", package = "IDclust")
scExp = iterative_differential_clustering(scExp,  saving = FALSE, plotting =FALSE,
logFC.th = 0.5, qval.th = 0.01)

}
#> 
#> 
#> 
#> 
#> Initial round of clustering - limit of differential genes set to 0  for this first round only.
#> ChromSCape::differential_activation - Calculating differential activation for A1 .
#> ChromSCape::differential_activation - Calculating differential activation for A2 .
#> A2  - found 962 enriched features.
#> A1  - found 954 enriched features.
#> Finished finding differences -  1  fraction of cells were assigned.
#> Doing iteration number  1 ...
#> Re-calculating PCA and subclustering for cluster Omega:A2 with 807 cells.
#> ChromScape::find_top_features - 5000 features kept as  most covered features...
#> ChromSCape::reduce_dims_scExp - removing Component_1 as probably correlated to library size.
#> 
#> 
#> 
#> 
#> Found 5 subclusters.
#> clusters
#>  B1  B2  B3  B4  B5 
#> 114 173 170 179 171 
#> ChromSCape::differential_activation - Calculating differential activation for B1 .
#> ChromSCape::differential_activation - Calculating differential activation for B2 .
#> ChromSCape::differential_activation - Calculating differential activation for B3 .
#> ChromSCape::differential_activation - Calculating differential activation for B4 .
#> ChromSCape::differential_activation - Calculating differential activation for B5 .
#> B1  - found 644 enriched features.
#> B3  - found 669 enriched features.
#> B5  - found 713 enriched features.
#> B2  - found 0 enriched features.
#> B2  cluster has less than 5 enriched features.
#> Assigning the cells to cluster of origin.
#> B4  - found 52 enriched features.
#> Finished finding differences -  0.7856258  fraction of cells were assigned.
#> Doing iteration number  2 ...
#> Re-calculating PCA and subclustering for cluster Omega:A1 with 1193 cells.
#> ChromScape::find_top_features - 5000 features kept as  most covered features...
#> ChromSCape::reduce_dims_scExp - removing Component_1 as probably correlated to library size.
#> 
#> 
#> 
#> 
#> Found 6 subclusters.
#> clusters
#>  B1  B2  B3  B4  B5  B6 
#> 349 209 155 181 168 131 
#> ChromSCape::differential_activation - Calculating differential activation for B1 .
#> ChromSCape::differential_activation - Calculating differential activation for B2 .
#> ChromSCape::differential_activation - Calculating differential activation for B3 .
#> ChromSCape::differential_activation - Calculating differential activation for B4 .
#> ChromSCape::differential_activation - Calculating differential activation for B5 .
#> ChromSCape::differential_activation - Calculating differential activation for B6 .
#> B6  - found 237 enriched features.
#> B3  - found 439 enriched features.
#> B5  - found 287 enriched features.
#> B4  - found 300 enriched features.
#> B2  - found 0 enriched features.
#> B2  cluster has less than 5 enriched features.
#> Assigning the cells to cluster of origin.
#> B1  - found 70 enriched features.
#> Finished finding differences -  0.8248114  fraction of cells were assigned.
#> Doing iteration number  3 ...
#> Re-calculating PCA and subclustering for cluster Omega:A2:B1 with 114 cells.
#> ChromScape::find_top_features - 5000 features kept as  most covered features...
#> ChromSCape::reduce_dims_scExp - removing Component_1 as probably correlated to library size.
#> 
#> 
#> 
#> 
#> Found only 1 subcluster for Omega:A2:B1 . Maximum clustering reached...
#> Doing iteration number  4 ...
#> Re-calculating PCA and subclustering for cluster Omega:A2:B3 with 170 cells.
#> ChromScape::find_top_features - 5000 features kept as  most covered features...
#> ChromSCape::reduce_dims_scExp - removing Component_1 as probably correlated to library size.
#> 
#> 
#> 
#> 
#> Found only 1 subcluster for Omega:A2:B3 . Maximum clustering reached...
#> Doing iteration number  5 ...
#> Re-calculating PCA and subclustering for cluster Omega:A2:B5 with 171 cells.
#> ChromScape::find_top_features - 5000 features kept as  most covered features...
#> ChromSCape::reduce_dims_scExp - removing Component_1 as probably correlated to library size.
#> 
#> 
#> 
#> 
#> Found only 1 subcluster for Omega:A2:B5 . Maximum clustering reached...
#> Doing iteration number  6 ...
#> Omega:A2 - This cluster was formed by 'unassigned' cells, not clustering it further...
#> Doing iteration number  7 ...
#> Re-calculating PCA and subclustering for cluster Omega:A2:B4 with 179 cells.
#> ChromScape::find_top_features - 5000 features kept as  most covered features...
#> ChromSCape::reduce_dims_scExp - removing Component_1 as probably correlated to library size.
#> 
#> 
#> 
#> 
#> Found 2 subclusters.
#> clusters
#>  C1  C2 
#> 146  33 
#> ChromSCape::differential_activation - Calculating differential activation for C1 .
#> ChromSCape::differential_activation - Calculating differential activation for C2 .
#> C2  - found 55 enriched features.
#> C1  - found 0 enriched features.
#> C1  cluster has less than 5 enriched features.
#> Assigning the cells to cluster of origin.
#> Finished finding differences -  0.1843575  fraction of cells were assigned.
#> Doing iteration number  8 ...
#> Re-calculating PCA and subclustering for cluster Omega:A1:B6 with 131 cells.
#> ChromScape::find_top_features - 5000 features kept as  most covered features...
#> ChromSCape::reduce_dims_scExp - removing Component_1 as probably correlated to library size.
#> 
#> 
#> 
#> 
#> Found only 1 subcluster for Omega:A1:B6 . Maximum clustering reached...
#> Doing iteration number  9 ...
#> Re-calculating PCA and subclustering for cluster Omega:A1:B3 with 155 cells.
#> ChromScape::find_top_features - 5000 features kept as  most covered features...
#> ChromSCape::reduce_dims_scExp - removing Component_1 as probably correlated to library size.
#> 
#> 
#> 
#> 
#> Found only 1 subcluster for Omega:A1:B3 . Maximum clustering reached...
#> Doing iteration number  10 ...
#> Re-calculating PCA and subclustering for cluster Omega:A1:B5 with 168 cells.
#> ChromScape::find_top_features - 5000 features kept as  most covered features...
#> ChromSCape::reduce_dims_scExp - removing Component_1 as probably correlated to library size.
#> 
#> 
#> 
#> 
#> Found only 1 subcluster for Omega:A1:B5 . Maximum clustering reached...
#> Doing iteration number  11 ...
#> Re-calculating PCA and subclustering for cluster Omega:A1:B4 with 181 cells.
#> ChromScape::find_top_features - 5000 features kept as  most covered features...
#> ChromSCape::reduce_dims_scExp - removing Component_1 as probably correlated to library size.
#> 
#> 
#> 
#> 
#> Found 2 subclusters.
#> clusters
#>  C1  C2 
#> 131  50 
#> ChromSCape::differential_activation - Calculating differential activation for C1 .
#> ChromSCape::differential_activation - Calculating differential activation for C2 .
#> C2  - found 22 enriched features.
#> C1  - found 0 enriched features.
#> C1  cluster has less than 5 enriched features.
#> Assigning the cells to cluster of origin.
#> Finished finding differences -  0.2762431  fraction of cells were assigned.
#> Doing iteration number  12 ...
#> Omega:A1 - This cluster was formed by 'unassigned' cells, not clustering it further...
#> Doing iteration number  13 ...
#> Re-calculating PCA and subclustering for cluster Omega:A1:B1 with 349 cells.
#> ChromScape::find_top_features - 5000 features kept as  most covered features...
#> ChromSCape::reduce_dims_scExp - removing Component_1 as probably correlated to library size.
#> 
#> 
#> 
#> 
#> Found 2 subclusters.
#> clusters
#>  C1  C2 
#> 207 142 
#> ChromSCape::differential_activation - Calculating differential activation for C1 .
#> ChromSCape::differential_activation - Calculating differential activation for C2 .
#> C2  - found 16 enriched features.
#> C1  - found 0 enriched features.
#> C1  cluster has less than 5 enriched features.
#> Assigning the cells to cluster of origin.
#> Finished finding differences -  0.4068768  fraction of cells were assigned.
#> Doing iteration number  14 ...
#> Re-calculating PCA and subclustering for cluster Omega:A2:B4:C2 with 33 cells.
#> Doing iteration number  15 ...
#> Omega:A2:B4 - This cluster was formed by 'unassigned' cells, not clustering it further...
#> Doing iteration number  16 ...
#> Re-calculating PCA and subclustering for cluster Omega:A1:B4:C2 with 50 cells.
#> Doing iteration number  17 ...
#> Omega:A1:B4 - This cluster was formed by 'unassigned' cells, not clustering it further...
#> Doing iteration number  18 ...
#> Re-calculating PCA and subclustering for cluster Omega:A1:B1:C2 with 142 cells.
#> ChromScape::find_top_features - 5000 features kept as  most covered features...
#> ChromSCape::reduce_dims_scExp - removing Component_1 as probably correlated to library size.
#> 
#> 
#> 
#> 
#> Found only 1 subcluster for Omega:A1:B1:C2 . Maximum clustering reached...
#> Doing iteration number  19 ...
#> Omega:A1:B1 - This cluster was formed by 'unassigned' cells, not clustering it further...
#> 
#> 
#> 
#> ##########################################################
#> Finished !
#> Found a total of 14 clusters after 19 iterations. 
#> The average cluster size is  142  and the median is 150 . 
#> The number of initital clusters not subclustered is  0 . 
#> ##########################################################